+++ WineCore/truth/handlers/userwine.py
@@ -1,11 +1,11 @@
 from truth.stubs import webapp2, ndb
-from google.appengine.api import users
 from truth.views.jsonview import json_response 
 from truth.models.event import Event
 from truth.models.winery import Winery
 from truth.models.wine import Wine
 from truth.models.userwine import UserWine
 from truth.constants import MAX_RESULTS
+from truth.models.user import User
 
 
 class UserWineBaseHandler(webapp2.RequestHandler):
@@ -32,7 +32,7 @@ class UserWineBaseHandler(webapp2.RequestHandler):
         userwine = UserWine(parent=wine_key)
 
         try:
-            post['user'] = users.get_current_user()
+            post['user'] = User.get_current_user()
             key = userwine.create(post)
             Event.create(self.request.remote_addr, "UserWine", key)
         except ValueError as e:
@@ -52,6 +52,11 @@ class UserWineHandler(webapp2.RequestHandler):
             self.response.status = "404 Not Found"
             return
 
+        if userwine.user != User.get_current_user().key:
+            self.response.write("403 Forbidden")
+            self.response.status = "403 Forbidden"            
+            return
+
         json_response(self, userwine)
 
     def post(self, winery_id, wine_id, userwine_id):
@@ -59,6 +64,16 @@ class UserWineHandler(webapp2.RequestHandler):
 
         userwine_key = ndb.Key(Winery, int(winery_id), Wine, int(wine_id), UserWine, int(userwine_id))
         userwine = userwine_key.get()
+        if not userwine:
+            self.response.write("404 Not Found")
+            self.response.status = "404 Not Found"
+            return
+
+        if userwine.user != User.get_current_user().key:
+            self.response.write("403 Forbidden")
+            self.response.status = "403 Forbidden"            
+            return
+
         userwine.modify(post)
         Event.update(self.request.remote_addr, "UserWine", userwine_key)
 
@@ -67,6 +82,16 @@ class UserWineHandler(webapp2.RequestHandler):
     def delete(self, winery_id, wine_id, userwine_id):
         userwine_key = ndb.Key(Winery, int(winery_id), Wine, int(wine_id), UserWine, int(userwine_id))
         userwine = userwine_key.get()
+        if not userwine:
+            self.response.write("404 Not Found")
+            self.response.status = "404 Not Found"
+            return
+
+        if userwine.user != User.get_current_user().key:
+            self.response.write("403 Forbidden")
+            self.response.status = "403 Forbidden"            
+            return
+
         userwine.delete()
         json_response(self, {"success":True})
