+++ WineCore/truth/handlers/winebottle.py
@@ -6,6 +6,7 @@ from truth.models.wine import Wine
 from truth.models.winery import Winery
 from truth.models.cellar import WineCellar
 from truth.models.event import Event
+from truth.models.user import User
 
 class WineBottleBaseHandler(webapp2.RequestHandler):
     def get(self, cellar_id):
@@ -15,6 +16,11 @@ class WineBottleBaseHandler(webapp2.RequestHandler):
         if not cellar:
             self.response.write("404 Not Found")
             self.response.status = "404 Not Found"
+            return
+
+        if not User.hasAccess(cellar_key):
+            self.response.write("403 Forbidden")
+            self.response.status = "403 Forbidden"            
             return
 
         qry = WineBottle.query(ancestor=cellar.key)        
@@ -30,6 +36,11 @@ class WineBottleBaseHandler(webapp2.RequestHandler):
         if not cellar:
             self.response.write("404 Not Found")
             self.response.status = "404 Not Found"
+            return
+
+        if not User.hasAccess(cellar_key):
+            self.response.write("403 Forbidden")
+            self.response.status = "403 Forbidden"            
             return
 
         post = self.request.POST
@@ -66,6 +77,11 @@ class WineBottleHandler(webapp2.RequestHandler):
             self.response.status = "404 Not Found"
             return
 
+        if not User.hasAccess(ndb.Key(WineCellar, int(cellar_id))):
+            self.response.write("403 Forbidden")
+            self.response.status = "403 Forbidden"            
+            return
+
         json_response(self, bottle)
 
     def post(self,cellar_id,winebottle_id):
@@ -75,6 +91,11 @@ class WineBottleHandler(webapp2.RequestHandler):
         if not bottle:
             self.response.write("404 Not Found")
             self.response.status = "404 Not Found"
+            return
+
+        if not User.hasAccess(ndb.Key(WineCellar, int(cellar_id))):
+            self.response.write("403 Forbidden")
+            self.response.status = "403 Forbidden"            
             return
 
         post = self.request.POST
@@ -108,7 +129,12 @@ class WineBottleHandler(webapp2.RequestHandler):
             self.response.write("404 Not Found")
             self.response.status = "404 Not Found"
             return
-        
+
+        if not User.hasAccess(ndb.Key(WineCellar, int(cellar_id))):
+            self.response.write("403 Forbidden")
+            self.response.status = "403 Forbidden"            
+            return
+                    
         bottle.delete()
 
         Event.create(self.request.remote_addr, "WineBottle", bottle_key)
